refactor : change js file to ts and using next-connect for api routing
