+++ package.json
@@ -17,6 +17,7 @@
     "dotenv": "^16.0.2",
     "mongoose": "^6.6.1",
     "next": "12.2.5",
+    "next-connect": "^0.13.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-query": "^3.39.2",
