+++ src/middlewares/connectMongo.ts
@@ -1,0 +1,4 @@
+import mongoose from 'mongoose';
+
+export const connectMongo = async () =>
+  await mongoose.connect(`${process.env.MONGO_URI}`);
