+++ src/models/articleModel.ts
