--- src/utils/connectMongo.js
@@ -1,3 +1,0 @@
-import mongoose from 'mongoose';
-
-export const connectMongo = async () => mongoose.connect(process.env.MONGO_URI);
